@@ -1,17 +1,18 @@
 name: authentication_repository
 description: Dart package which manages the authentication domain.
-version: 1.0.0
+version: 1.1.0
 publish_to: none
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ^3.4.0
+  flutter: ^3.24.1
 
 dependencies:
   cache:
     path: ../cache
   equatable: ^2.0.5
-  firebase_auth: ^4.14.1
-  firebase_core: ^2.23.0
+  firebase_auth: ^5.2.0
+  firebase_core: ^3.4.0
   flutter:
     sdk: flutter
   google_sign_in: ^6.1.0
