@@ -1,10 +1,11 @@
 name: cache
 description: A simple in memory cache made for dart
-version: 1.0.0
+version: 1.1.0
 publish_to: none
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ^3.4.0
+  flutter: ^3.24.1
 
 dev_dependencies:
   test: ^1.17.11
