@@ -1,9 +1,10 @@
 name: form_inputs
-version: 1.0.0+1
+version: 1.1.0+1
 publish_to: none
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ^3.4.0
+  flutter: ^3.24.1
 
 dependencies:
   formz: ^0.6.0
