@@ -1,35 +1,34 @@
 name: bbnaf
 description: Bloodbowl tournament management App.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
-version: 2.1.0+1
+version: 3.1.0+1
 
 environment:
-  sdk: ">=3.2.0 <4.0.0"
+  sdk: ^3.4.0
+  flutter: ^3.24.1
 
 dependencies:
-  adaptive_dialog: ^2.0.0
+  adaptive_dialog: ^2.2.0
   authentication_repository:
     path: packages/authentication_repository
   bloc: ^8.1.0
   bloc_concurrency: ^0.2.3
   cache:
     path: packages/cache
-  checkbox_formfield: ^0.2.0
-  cloud_firestore: ^4.3.2
+  cloud_firestore: ^5.3.0
   cupertino_icons: ^1.0.0
   data_tables: ^1.0.3
   data_table_2: ^2.5.10
-  easy_search_bar: ^2.5.0
   english_words: ^4.0.0
   enum_to_string: ^2.0.1
   equatable: ^2.0.3
-  excel: ^4.0.2
+  excel: ^4.0.6
   file_picker: ^6.1.1
-  firebase_auth: ^4.2.6
-  firebase_core: ^2.5.0
-  firebase_messaging: ^14.2.4
-  firebase_storage: ^11.0.11
-  firebase_storage_web: ^3.3.20
+  firebase_auth: ^5.2.0
+  firebase_core: ^3.4.0
+  firebase_messaging: ^15.1.0
+  firebase_storage: ^12.2.0
+  firebase_storage_web: ^3.9.13
   flow_builder: ^0.0.9
   flutter:
     sdk: flutter
@@ -38,7 +37,7 @@ dependencies:
   flutter_native_splash: ^2.3.9
   flutter_rating_bar: ^4.0.1
   flutter_widget_from_html: ^0.14.11
-  flutter_quill: ^9.2.12
+  flutter_quill: ^10.0.7
   fluttertoast: ^8.1.4
   font_awesome_flutter: ^10.1.0
   form_inputs:
@@ -46,16 +45,17 @@ dependencies:
   formz: ^0.6.0
   google_fonts: ^4.0.0
   grouped_list: ^5.1.2
-  intl: ^0.18.1
+  intl: ^0.19.0
   json_annotation: ^4.0.1
   json_serializable: ^6.7.1
   meta: ^1.7.0
   responsive_framework: ^1.1.1 
-  screenshot: ^2.3.0
+  screenshot: ^3.0.0
   string_similarity: ^2.0.0
   syncfusion_flutter_datepicker: ^24.1.47
-  uni_links: ^0.5.1
+  # uni_links: ^0.5.1 replaced by app_links
   uuid: ^4.0.0
+  web: ^0.5.1
   xml: ^6.5.0
 
 
