@@ -7,7 +7,8 @@ import 'package:bbnaf/matchups/matchups.dart';
 import 'package:bbnaf/home/view/overview_screen.dart';
 import 'package:bbnaf/utils/buy_me_a_coffee/buy_me_a_coffee.dart';
 import 'package:bbnaf/utils/serializable.dart';
-import 'package:easy_search_bar/easy_search_bar.dart';
+import 'package:bbnaf/widgets/easy_search_bar/easy_search_bar.dart';
+// import 'package:easy_search_bar/easy_search_bar.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:google_fonts/google_fonts.dart';
