@@ -76,7 +76,7 @@ class BorderedText extends StatelessWidget {
           strutStyle: child.strutStyle,
           textAlign: child.textAlign,
           textDirection: child.textDirection,
-          textScaleFactor: child.textScaleFactor,
+          textScaler: child.textScaler,
         ),
         child,
       ],
